pyproject.toml
[project]
name = "anibridge-mappings"
version = "3.0.0"
description = "Anime mappings for the AniBridge project."
readme = "README.md"
requires-python = ">=3.14"
license = { text = "MIT" }

authors = [
    { name = "Elias Benbourenane", email = "eliasbenbourenane@gmail.com" }
]
maintainers = [
    { name = "Elias Benbourenane", email = "eliasbenbourenane@gmail.com" },
    { name = "LuceoEtzio" }
]

dependencies = [
    "anibridge-utils>=0.2.0b3",
    "aiohttp>=3.13.2",
    "anibridge-utils>=0.1.1",
    "dotenv>=0.9.9",
    "lxml>=6.0.2",
    "pydantic>=2.12.5",
    "ruamel-yaml>=0.19.1",
    "zstandard>=0.25.0",
]

[project.urls]
Homepage = "https://anibridge.eliasbenb.dev"
Documentation = "https://anibridge.eliasbenb.dev"
Source = "https://github.com/anibridge/anibridge-mappings"
Issues = "https://github.com/anibridge/anibridge-mappings/issues"

[tool.ruff]
indent-width = 4
line-length = 88

[tool.ruff.format]
docstring-code-format = true
indent-style = "space"
quote-style = "double"

[tool.ruff.lint]
select = ["B", "D", "DOC", "E", "F", "I", "RUF", "SIM", "UP", "W"]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["D"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]
addopts = "--cov=src"
filterwarnings = []

[build-system]
requires = ["uv_build>=0.10.9,<0.11.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "pytest>=9.0.2",
    "pytest-cov>=7.1.0",
    "ruff>=0.15.5",
    "ty>=0.0.22",
    "types-lxml>=2025.11.25",
]
